+++ src/clues/avoid-emoji/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { checkEmoji } from "../avoid-emoji";
+import { checkEmoji } from "..";
 
 test("checkEmoji", () => {
   expect(checkEmoji("An orange 🐈.")).toMatchInlineSnapshot(`
