+++ src/clues/avoid-emoji/index.ts
@@ -1,7 +1,7 @@
 import emojiRegex from "emoji-regex";
-import { Rule } from "./index";
-import { createSuggestion } from "../check-clue";
-import { Alt } from "../index.js";
+import { Rule } from "../index";
+import { createSuggestion } from "../../check-clue";
+import { Alt } from "../../index.js";
 
 export function checkEmoji(alt: Alt) {
   const regex = emojiRegex();
