+++ src/clues/char-length/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { checkLength } from "../char-length";
+import { checkLength } from "..";
 
 test("checkLength", () => {
   expect(
