+++ src/clues/contains/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import checkClue from "../../check-clue";
+import checkClue from "../../../check-clue";
 
 test("contains", () => {
   expect(checkClue("A screenshot of a dog.")).toMatchInlineSnapshot(`
