+++ src/clues/end-punctuation/index.ts
@@ -1,6 +1,6 @@
-import { Alt } from "../index";
-import { Rule } from "./index";
-import { createSuggestion } from "../check-clue";
+import { Alt } from "../../index";
+import { Rule } from "../index";
+import { createSuggestion } from "../../check-clue";
 
 export function checkPunctuation(alt: Alt) {
   return !/[.!?]$/.test(alt) && alt.length > 1
