+++ src/clues/image-link/index.ts
@@ -1,4 +1,4 @@
-import { Rule } from "./index";
+import { Rule } from "../index";
 
 export default function imageLink(): Rule {
   return {
