+++ src/clues/not-only-space/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { checkOnlySpace } from "../not-only-space";
+import { checkOnlySpace } from "..";
 
 test("checkOnlySpace", () => {
   expect(checkOnlySpace(" ")).toMatchInlineSnapshot(`
