+++ src/clues/start-with/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import checkClue from "../../check-clue";
+import checkClue from "../../../check-clue";
 
 test("startWith", () => {
   expect(checkClue("spacer image.")).toMatchInlineSnapshot(`
